+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Authentication - Customisable expiration time to the liana token.
 
 ## RELEASE 2.0.0 - 2017-11-24
 ### Added